refactor(web-providers): split content-fetch-main.ts into per-provider files
Extract 6 provider-specific MAIN world handlers from the 3,756-line
content-fetch-main.ts monolith into separate files. The bridge now
injects only the needed handler function via chrome.scripting.executeScript,
reducing the per-request injection payload.

New files:
- content-fetch-gemini.ts (187 lines)
- content-fetch-glm-intl.ts (356 lines)
- content-fetch-deepseek.ts (623 lines)
- content-fetch-doubao.ts (238 lines)
- content-fetch-chatgpt.ts (1,028 lines)
- content-fetch-rakuten.ts (1,033 lines)

content-fetch-main.ts shrinks to 331 lines (shared Connect Protocol +
standard SSE streaming used by claude-web, qwen-web, qwen-cn-web,
glm-web, and kimi-web).

Each handler is wired into its plugin via contentFetchHandler. The bridge
falls back to mainWorldFetch for providers without a custom handler.
