+++ chrome-extension/src/background/web-providers/providers/chatgpt-web.plugin.ts
@@ -2,11 +2,13 @@ import type { WebProviderPlugin } from '../plugin-types';
 import { chatgptWeb } from './chatgpt-web';
 import { createChatGPTStreamAdapter } from './chatgpt-stream-adapter';
 import { chatgptToolStrategy } from '../tool-strategy';
+import { chatgptMainWorldFetch } from '../content-fetch-chatgpt';
 
 export const chatgptWebPlugin: WebProviderPlugin = {
   definition: chatgptWeb,
   createStreamAdapter: () => createChatGPTStreamAdapter(),
   toolStrategy: chatgptToolStrategy,
+  contentFetchHandler: chatgptMainWorldFetch,
   hooks: {
     onStreamDone: ({ credential }) => {
       // Dynamic import to avoid pulling chrome.storage into test environments
